--- src/main/java/roomescape/repository/ReservationThemeDao.java
@@ -1,18 +1,0 @@
-package roomescape.repository;
-
-import java.util.List;
-import java.util.Optional;
-import roomescape.domain.ReservationTheme;
-
-public interface ReservationThemeDao {
-
-    ReservationTheme save(ReservationTheme reservationTheme);
-
-    List<ReservationTheme> findAll();
-
-    void delete(Long id);
-
-    Optional<ReservationTheme> findById(Long id);
-
-    Long findByName(String name);
-}
